cmd/atlas/internal/cmdapi: allow extending schema diff with --include
