+++ cmd/atlas/internal/cmdapi/schema.go
@@ -14,11 +14,9 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"text/template"
 	"time"
 
 	"ariga.io/atlas/cmd/atlas/internal/cmdext"
-	"ariga.io/atlas/cmd/atlas/internal/cmdlog"
 	"ariga.io/atlas/sql/migrate"
 	"ariga.io/atlas/sql/schema"
 	"ariga.io/atlas/sql/sqlclient"
@@ -245,6 +243,7 @@ type schemaDiffFlags struct {
 	devURL  string
 	schemas []string
 	exclude []string
+	include []string
 	format  string
 }
 
@@ -285,67 +284,11 @@ The database states can be read from a connected database, an HCL project or a m
 	addFlagDevURL(cmd.Flags(), &flags.devURL)
 	addFlagSchemas(cmd.Flags(), &flags.schemas)
 	addFlagExclude(cmd.Flags(), &flags.exclude)
+	addFlagInclude(cmd.Flags(), &flags.include)
 	addFlagFormat(cmd.Flags(), &flags.format)
 	cobra.CheckErr(cmd.MarkFlagRequired(flagFrom))
 	cobra.CheckErr(cmd.MarkFlagRequired(flagTo))
 	return cmd, &flags
-}
-
-func schemaDiffRun(cmd *cobra.Command, _ []string, flags schemaDiffFlags, env *Env) error {
-	var (
-		ctx = cmd.Context()
-		c   *sqlclient.Client
-	)
-	// We need a driver for diffing and planning. If given, dev database has precedence.
-	if flags.devURL != "" {
-		var err error
-		c, err = sqlclient.Open(ctx, flags.devURL)
-		if err != nil {
-			return err
-		}
-		defer c.Close()
-	}
-	from, err := stateReader(ctx, env, &stateReaderConfig{
-		urls:    flags.fromURL,
-		dev:     c,
-		vars:    env.Vars(),
-		schemas: flags.schemas,
-		exclude: flags.exclude,
-	})
-	if err != nil {
-		return err
-	}
-	defer from.Close()
-	to, err := stateReader(ctx, env, &stateReaderConfig{
-		urls:    flags.toURL,
-		dev:     c,
-		vars:    env.Vars(),
-		schemas: flags.schemas,
-		exclude: flags.exclude,
-	})
-	if err != nil {
-		return err
-	}
-	defer to.Close()
-	if c == nil {
-		// If not both states are provided by a database connection, the call to state-reader would have returned
-		// an error already. If we land in this case, we can assume both states are database connections.
-		c = to.Closer.(*sqlclient.Client)
-	}
-	format := cmdlog.SchemaDiffTemplate
-	if v := flags.format; v != "" {
-		if format, err = template.New("format").Funcs(cmdlog.SchemaDiffFuncs).Parse(v); err != nil {
-			return fmt.Errorf("parse log format: %w", err)
-		}
-	}
-	diff, err := computeDiff(ctx, c, from, to, diffOptions(cmd, env)...)
-	if err != nil {
-		return err
-	}
-	maySuggestUpgrade(cmd)
-	return format.Execute(cmd.OutOrStdout(),
-		cmdlog.NewSchemaDiff(ctx, c, diff.from, diff.to, diff.changes),
-	)
 }
 
 // schemaFmtCmd represents the 'atlas schema fmt' subcommand.
